change products

+++ src/main/kotlin/com/utsman/repository/CategoryRepository.kt
@@ -9,7 +9,7 @@ class CategoryRepository {
 
     suspend fun readCategoriesCsv(): List<Category> {
         return readCsvList("categories.csv") { fields ->
-            val id = fields[0].trim()
+            val id = fields[0].trim().toInt()
             val name = fields[1].trim()
             val description = fields[2].trim()
             Category(id, name, description)
